Extract telnet protocol constants into telnetsrv/constants.py
Moves all protocol-level constants (IAC, DO, WILL, NOOPT, CMDS, etc.)
and ANSI/curses helper tables out of telnetsrvlib.py into a dedicated
constants module. telnetsrvlib re-exports them via star import for
backward compatibility; aio.py and all test files now import constants
directly from telnetsrv.constants.

Also makes the 0.5s option-negotiation delay configurable via the
OPTION_NEGOTIATION_DELAY class attribute, and adds a README note
clarifying that write methods are synchronous and must not be awaited.
